feat : structure updated
